requirements.txt
anyio==4.12.0
cachetools==6.2.3
certifi==2025.11.12
charset-normalizer==3.4.4
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
openaq==0.6.0
packaging==25.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.2
